Update to JFreeChart 1.5.0
Large amount of tweaking to plot settings was necessary to match legacy
appearance, due to wide changes to defaults.
Updated JCommon as well, but we may actually be able to remove this
entirely (supposedly no longer needed by JFreeChart).  That will be
evaluated later.
ProGuard is not fully working yet after the update.